transmission/requirements.txt
PyYAML==6.0.1
redis==5.0.1
transmission-rpc==7.0.3
